requirements.txt
PyPDF2
pdfplumber==0.11.5
python-docx==1.1.2
sentence-transformers==3.4.1
numpy==1.26.4
nltk==3.9.1
tf-keras==2.18.0
tensorflow==2.18.0
pymilvus==2.5.4
streamlit==1.42.0
openai==1.61.1